atualiza projetos

+++ Chapter_7/src/tryOutPlaceToLive/PlaceToLive.java
@@ -1,5 +1,9 @@
 package tryOutPlaceToLive;
 
+import java.util.Scanner;
+
+import static java.lang.System.out;
+
 public class PlaceToLive {
 
     /*204 PART 3 Working with the Big Picture: Object-Oriented Programming
@@ -7,13 +11,18 @@ public class PlaceToLive {
 
         » A PlaceToLive has an address, a number of bedrooms, and an area (in
         square feet or square meters).
-        Write the PlaceToLive class’s code. Write code
-        for a separate class named DisplayThePlaces.
+
+        Write the PlaceToLive class’s code.
+        Write code for a separate class named DisplayThePlaces.
+
         Your DisplayThePlaces class creates a few PlaceToLive instances by assigning values to their
         address, numberOfBedrooms, and area fields.
+
         The DisplayThePlaces class also reads (from the keyboard) the cost of living in each place.
+
         For each place,your code displays the cost per square foot (or square meter) and the cost
         per bedroom.
+
         » Use your new PlaceToLive class and my DummiesFrame class (from
         Chapter 7) to create a GUI application. The GUI application takes information
         about a place to live and displays the place’s cost per square foot (or meter)
@@ -21,7 +30,7 @@ public class PlaceToLive {
 
     private String address;
     private int numberBedrooms;
-    private double sizeArea;
+    private int sizeArea;
 
     public String getAddress() {
         return address;
@@ -39,15 +48,18 @@ public class PlaceToLive {
         this.numberBedrooms = numberBedrooms;
     }
 
-    public double getSizeArea() {
+    public int getSizeArea() {
         return sizeArea;
     }
 
-    public void setSizeArea(double sizeArea) {
+    public void setSizeArea(int sizeArea) {
         this.sizeArea = sizeArea;
     }
 
     // calcula o valor do metro quadrado
+
+
+
 
 
 }
